+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/utils.py
@@ -58,12 +58,9 @@ def timestamp_to_iso_str(timestamp):
     return to_iso_str(datetime.datetime.utcfromtimestamp(timestamp))
 
 
-# Validate UUID format
-# Specs taken from https://tools.ietf.org/html/rfc4122
+# Validate GUID format
 uuid_regex_pattern = re.compile('^[0-9a-f]{8}-'
-                                '[0-9a-f]{4}-'
-                                '[1-5][0-9a-f]{3}-'
-                                '[89ab][0-9a-f]{3}-'
+                                '([0-9a-f]{4}-){3}'
                                 '[0-9a-f]{12}$')