+++ contrib/opencensus-ext-azure/opencensus/ext/azure/metrics_exporter/__init__.py
@@ -153,9 +153,4 @@ def new_metrics_exporter(**options):
                                     producers,
                                     exporter,
                                     interval=exporter.options.export_interval)
-    from opencensus.ext.azure.metrics_exporter import heartbeat_metrics
-    heartbeat_metrics.enable_heartbeat_metrics(
-        exporter.options.connection_string,
-        exporter.options.instrumentation_key
-    )
     return exporter